refactor query engine
