+++ query/contracts/query_contracts_test.go
@@ -1,10 +1,11 @@
 package query
 
 import (
-	"canto-api/config"
-	"canto-api/multicall"
 	"reflect"
 	"testing"
+
+	"canto-api/config"
+	"canto-api/multicall"
 )
 
 func TestProcessContractCalls(t *testing.T) {
