+++ query/contracts/utils_test.go
@@ -1,11 +1,12 @@
 package query
 
 import (
-	"canto-api/multicall"
 	"encoding/hex"
 	"fmt"
 	"reflect"
 	"testing"
+
+	"canto-api/multicall"
 
 	"github.com/ethereum/go-ethereum/common"
 )
